refactor(config): move provider capability metadata to providers/capabilities.py
Move default models, provider env var chains, base URLs, and embedding
dimension rules from config.py into vexor/providers/capabilities.py per
the roadmap Engineering TODO. config.py permanently re-exports every
moved name, so the public Python API is unchanged. Runtime-internal
imports now target the new module; tests keep importing from
vexor.config as re-export regression coverage.
