+++ vexor/config.py
@@ -12,6 +12,27 @@ from pathlib import Path
 from typing import Any, Dict
 from urllib.parse import urlparse, urlunparse
 
+# Permanent public re-exports preserve existing imports from vexor.config.
+from .providers.capabilities import (
+    DEFAULT_GEMINI_MODEL,
+    DEFAULT_LOCAL_MODEL,
+    DEFAULT_MODEL,
+    DEFAULT_PROVIDER,
+    DEFAULT_VOYAGE_MODEL,
+    DIMENSION_SUPPORTED_MODELS,
+    ENV_API_KEY,
+    LEGACY_GEMINI_ENV,
+    OPENAI_ENV,
+    SUPPORTED_PROVIDERS,
+    VOYAGE_BASE_URL,
+    VOYAGE_ENV,
+    get_supported_dimensions,
+    resolve_api_key,
+    resolve_base_url,
+    resolve_default_model,
+    supports_dimensions,
+    validate_embedding_dimensions_for_model,
+)
 from .text import Messages
 
 DEFAULT_CONFIG_DIR = Path(os.path.expanduser("~")) / ".vexor"
@@ -21,36 +42,18 @@ _CONFIG_DIR_OVERRIDE: ContextVar[Path | None] = ContextVar(
     "vexor_config_dir_override",
     default=None,
 )
-DEFAULT_MODEL = "text-embedding-3-small"
-DEFAULT_GEMINI_MODEL = "gemini-embedding-001"
-DEFAULT_VOYAGE_MODEL = "voyage-3-large"
-DEFAULT_LOCAL_MODEL = "intfloat/multilingual-e5-small"
 DEFAULT_BATCH_SIZE = 64
 DEFAULT_EMBED_CONCURRENCY = 4
 DEFAULT_EXTRACT_CONCURRENCY = max(1, min(4, os.cpu_count() or 1))
 DEFAULT_EXTRACT_BACKEND = "auto"
-DEFAULT_PROVIDER = "openai"
 DEFAULT_RERANK = "off"
 DEFAULT_FLASHRANK_MODEL = "ms-marco-TinyBERT-L-2-v2"
 DEFAULT_FLASHRANK_MAX_LENGTH = 256
-VOYAGE_BASE_URL = "https://api.voyageai.com/v1"
-SUPPORTED_PROVIDERS: tuple[str, ...] = (DEFAULT_PROVIDER, "gemini", "voyageai", "custom", "local")
 SUPPORTED_RERANKERS: tuple[str, ...] = ("off", "bm25", "flashrank", "remote")
 SUPPORTED_EXTRACT_BACKENDS: tuple[str, ...] = ("auto", "thread", "process")
-# Models that support the dimensions parameter (model prefix/name -> supported dimensions)
-DIMENSION_SUPPORTED_MODELS: dict[str, tuple[int, ...]] = {
-    "text-embedding-3-small": (256, 512, 1024, 1536),
-    "text-embedding-3-large": (256, 512, 1024, 1536, 3072),
-    "voyage-3": (256, 512, 1024, 2048),
-    "voyage-code-3": (256, 512, 1024, 2048),
-}
-ENV_API_KEY = "VEXOR_API_KEY"
 ENV_CONFIG_JSON = "VEXOR_CONFIG_JSON"
 ENV_NO_UPDATE_CHECK = "VEXOR_NO_UPDATE_CHECK"
 REMOTE_RERANK_ENV = "VEXOR_REMOTE_RERANK_API_KEY"
-LEGACY_GEMINI_ENV = "GOOGLE_GENAI_API_KEY"
-OPENAI_ENV = "OPENAI_API_KEY"
-VOYAGE_ENV = "VOYAGE_API_KEY"
 
 
 @dataclass
@@ -460,60 +463,6 @@ def normalize_remote_rerank_url(value: str | None) -> str | None:
     return urlunparse(normalized)
 
 
-def resolve_default_model(provider: str | None, model: str | None) -> str:
-    """Return the effective model name for the selected provider."""
-    clean_model = (model or "").strip()
-    normalized = (provider or DEFAULT_PROVIDER).lower()
-    if normalized == "gemini" and (not clean_model or clean_model == DEFAULT_MODEL):
-        return DEFAULT_GEMINI_MODEL
-    if normalized == "voyageai" and (not clean_model or clean_model == DEFAULT_MODEL):
-        return DEFAULT_VOYAGE_MODEL
-    if clean_model:
-        return clean_model
-    return DEFAULT_MODEL
-
-
-def resolve_base_url(provider: str | None, configured_url: str | None) -> str | None:
-    """Return the effective base URL for the selected provider."""
-    if configured_url:
-        return configured_url
-    normalized = (provider or DEFAULT_PROVIDER).lower()
-    if normalized == "voyageai":
-        return VOYAGE_BASE_URL
-    return None
-
-
-def supports_dimensions(model: str) -> bool:
-    """Check if a model supports the dimensions parameter."""
-    return get_supported_dimensions(model) is not None
-
-
-def get_supported_dimensions(model: str) -> tuple[int, ...] | None:
-    """Return the supported dimensions for a model, or None if not supported."""
-    model_lower = model.lower()
-    for prefix, dims in DIMENSION_SUPPORTED_MODELS.items():
-        if model_lower.startswith(prefix):
-            return dims
-    return None
-
-
-def validate_embedding_dimensions_for_model(value: int | None, model: str) -> None:
-    """Validate that `value` is supported by `model` when value is set."""
-    if value is None:
-        return
-    supported = get_supported_dimensions(model)
-    if not supported:
-        raise ValueError(
-            f"Model '{model}' does not support custom dimensions. "
-            f"Supported model names/prefixes: {', '.join(DIMENSION_SUPPORTED_MODELS.keys())}"
-        )
-    if value not in supported:
-        raise ValueError(
-            f"Dimension {value} is not supported for model '{model}'. "
-            f"Supported dimensions: {supported}"
-        )
-
-
 def _validate_config_embedding_dimensions(config: Config) -> None:
     """Ensure stored embedding dimensions remain compatible with provider/model."""
     if config.embedding_dimensions is None:
@@ -530,32 +479,6 @@ def _validate_config_embedding_dimensions(config: Config) -> None:
             f"model '{effective_model}'. Clear it with "
             "`vexor config --clear-embedding-dimensions` or set a supported value."
         ) from exc
-
-
-def resolve_api_key(configured: str | None, provider: str) -> str | None:
-    """Return the first available API key from config or environment."""
-
-    normalized = (provider or DEFAULT_PROVIDER).lower()
-    if normalized == "local":
-        return None
-    if configured:
-        return configured
-    general = os.getenv(ENV_API_KEY)
-    if general:
-        return general
-    if normalized == "gemini":
-        legacy = os.getenv(LEGACY_GEMINI_ENV)
-        if legacy:
-            return legacy
-    if normalized == "voyageai":
-        voyage_key = os.getenv(VOYAGE_ENV)
-        if voyage_key:
-            return voyage_key
-    if normalized in {"openai", "custom"}:
-        openai_key = os.getenv(OPENAI_ENV)
-        if openai_key:
-            return openai_key
-    return None
 
 
 def resolve_remote_rerank_api_key(configured: str | None) -> str | None:
